requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
autopep8==1.5.7
boto3==1.17.69
botocore==1.20.112
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
dj-database-url==0.5.0
Django==3.1.4
django-crispy-forms==1.10.0
django-crum==0.7.9
django-environ==0.4.5
django-extensions==3.1.2
django-filter==2.4.0
django-heroku==0.0.0
django-phonenumber-field==6.0.0
django-preventconcurrentlogins==0.8.2
django-resized==0.3.11
django-simple-history==2.12.0
django-storages==1.11.1
djangorestframework==3.12.4
drf-writable-nested==0.6.3
et-xmlfile==1.1.0
fuzzywuzzy==0.18.0
gunicorn==20.1.0
idna==3.2
jmespath==0.10.0
lml==0.1.0
ndg-httpsclient==0.5.1
openpyxl==3.0.7
Pillow==8.2.0
psycopg2-binary==2.9.1
pusher==3.0.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pyexcel-io==0.6.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
PyJWT==1.7.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
python-dateutil==2.8.2
python-decouple==3.4
python-Levenshtein==0.12.2
pytz==2020.4
requests==2.26.0
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
thefuzz==0.19.0
toml==0.10.2
twilio==6.52.0
urllib3==1.26.6
whitenoise==5.3.0
xlrd==1.2.0
xlwt==1.3.0
